Rename ModelizerNavigationBar to NavigationBar
