+++ src/components/NavigationBar.vue
